Update applemenu module to use only JDK APIs.
Merge adapters to only use JDK APIs.
Remove use of external binaries.
Update to release 17.
