--- platform/applemenu/external/orange-extensions-1.3.1-license.txt
@@ -1,39 +1,0 @@
-Name: Orange Extensions
-Version: 1.3.1
-License: BSD-Masory
-Description: com.apple classes for cross-compilation.
-Origin: Yuvi Masory
-Type: compile-time
-Comment: Allows to build code leveraging Apple-specficic classes on other platforms.
-
-This jar was created by decompiling AppleJavaExtensions.jar (completely consistent
-with its BSD-like license), and adding the new Java 5 methods.
-
-Apple's license does not specify that their copyright notice should be included
-in modified versions (only complete, unmodified ones), so I'm putting it under 3-clause
-BSD just to protect myself from liability.
-
-Copyright (c) 2010, Yuvi Masory
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name Yuvi Masory nor the
-      names of other contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL Yuvi Masory BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
