+++ platform/applemenu/nbproject/project.properties
@@ -16,9 +16,6 @@
 # under the License.
 
 javac.compilerargs=-Xlint -Xlint:-serial
-javac.source=1.8
+javac.release=17
 nbm.needs.restart=true
 is.eager=true
-cp.extra=external/orange-extensions-1.3.1.jar
-bootclasspath.prepend=${build.dir}/desktop-classes-classes
-
