+++ platform/applemenu/src/org/netbeans/modules/applemenu/NbApplicationAdapter.java
@@ -16,12 +16,21 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.netbeans.modules.applemenu;
 
+import java.awt.Desktop;
 import java.awt.Dialog;
 import java.awt.Frame;
 import java.awt.Window;
+import java.awt.desktop.AboutEvent;
+import java.awt.desktop.AboutHandler;
+import java.awt.desktop.OpenFilesEvent;
+import java.awt.desktop.OpenFilesHandler;
+import java.awt.desktop.PreferencesEvent;
+import java.awt.desktop.PreferencesHandler;
+import java.awt.desktop.QuitEvent;
+import java.awt.desktop.QuitHandler;
+import java.awt.desktop.QuitResponse;
 
 import java.awt.event.ActionEvent;
 import java.io.File;
@@ -44,18 +53,30 @@ import org.openide.windows.WindowManager;
 import org.openide.windows.WindowSystemEvent;
 import org.openide.windows.WindowSystemListener;
 
-/** Adapter class which intercepts action events and passes them to the
- * correct action instance as defined in the system filesystem.
+/**
+ * Adapter class which intercepts action events and passes them to the correct
+ * action instance as defined in the system filesystem.
  *
- * @author  Tim Boudreau
+ * @author Tim Boudreau
  */
+final class NbApplicationAdapter implements AboutHandler, OpenFilesHandler, PreferencesHandler, QuitHandler {
 
-abstract class NbApplicationAdapter {
-    
-    NbApplicationAdapter() {
+    private NbApplicationAdapter() {
     }
 
     static void install() {
+        try {
+            Desktop app = Desktop.getDesktop();
+            NbApplicationAdapter al = new NbApplicationAdapter();
+
+            app.setAboutHandler(al);
+            app.setOpenFileHandler(al);
+            app.setPreferencesHandler(al);
+            app.setQuitHandler(al);
+        } catch (Throwable ex) {
+            ErrorManager.getDefault().notify(ErrorManager.WARNING, ex);
+        }
+
         WindowManager.getDefault().addWindowSystemListener(new WindowSystemListener() {
 
             @Override
@@ -63,10 +84,10 @@ abstract class NbApplicationAdapter {
                 WindowManager.getDefault().removeWindowSystemListener(this);
                 try {
                     Frame main = WindowManager.getDefault().getMainWindow();
-                    ((RootPaneContainer)main).getRootPane().putClientProperty("apple.awt.fullscreenable", true);    // NOI18N
-                } catch( Throwable e ) {
-                    Logger.getLogger(NbApplicationAdapter.class.getName()).log(Level.FINE, 
-                            "Error while setting up full screen support.", e );//NOI18N
+                    ((RootPaneContainer) main).getRootPane().putClientProperty("apple.awt.fullscreenable", true);    // NOI18N
+                } catch (Throwable e) {
+                    Logger.getLogger(NbApplicationAdapter.class.getName()).log(Level.FINE,
+                            "Error while setting up full screen support.", e);//NOI18N
                 }
             }
 
@@ -84,15 +105,50 @@ abstract class NbApplicationAdapter {
         });
     }
 
-    void handleAbout() {
+    static void uninstall() {
+        try {
+            Desktop app = Desktop.getDesktop();
+
+            app.setAboutHandler(null);
+            app.setOpenFileHandler(null);
+            app.setPreferencesHandler(null);
+            app.setQuitHandler(null);
+        } catch (Throwable ex) {
+            ErrorManager.getDefault().notify(ErrorManager.WARNING, ex);
+        }
+    }
+
+    @Override
+    public void handleAbout(AboutEvent e) {
+        handleAbout();
+    }
+
+    @Override
+    public void openFiles(OpenFilesEvent e) {
+        openFiles(e.getFiles());
+    }
+
+    @Override
+    public void handlePreferences(PreferencesEvent e) {
+        handlePreferences();
+    }
+
+    @Override
+    public void handleQuitRequestWith(QuitEvent e, QuitResponse response) {
+        handleQuit();
+        //need to do this otherwise the user will never be able to quit again
+        response.cancelQuit();
+    }
+
+    private void handleAbout() {
         //#221571 - check if About window is showing already
         Window[] windows = Dialog.getWindows();
-        if( null != windows ) {
-            for( Window w : windows ) {
-                if( w instanceof JDialog ) {
+        if (null != windows) {
+            for (Window w : windows) {
+                if (w instanceof JDialog) {
                     JDialog dlg = (JDialog) w;
-                    if( Boolean.TRUE.equals(dlg.getRootPane().getClientProperty("nb.about.dialog") ) ) { //NOI18N
-                        if( dlg.isVisible() ) {
+                    if (Boolean.TRUE.equals(dlg.getRootPane().getClientProperty("nb.about.dialog"))) { //NOI18N
+                        if (dlg.isVisible()) {
                             dlg.toFront();
                             return;
                         }
@@ -102,15 +158,15 @@ abstract class NbApplicationAdapter {
         }
         performAction("Help", "org.netbeans.core.actions.AboutAction"); // NOI18N
     }
-    
-    void openFiles(List<File> files) {
+
+    private void openFiles(List<File> files) {
         for (File f : files) {
             if (f.exists() && !f.isDirectory()) {
                 FileObject obj = FileUtil.toFileObject(f);
                 if (obj != null) {
                     try {
                         DataObject dob = DataObject.find(obj);
-                        OpenCookie oc = dob.getLookup().lookup (OpenCookie.class);
+                        OpenCookie oc = dob.getLookup().lookup(OpenCookie.class);
                         if (oc != null) {
                             oc.open();
                         } else {
@@ -131,14 +187,15 @@ abstract class NbApplicationAdapter {
             }
         }
     }
-    
-    public void handlePreferences() {
+
+    private void handlePreferences() {
         performAction("Window", "org.netbeans.modules.options.OptionsWindowAction");    // NOI18N
     }
-    public void handleQuit() {
+
+    private void handleQuit() {
         performAction("System", "org.netbeans.core.actions.SystemExit");    // NOI18N
     }
-    
+
     private boolean performAction(String category, String id) {
         Action a = Actions.forID(category, id);
         if (a == null) {
@@ -153,5 +210,5 @@ abstract class NbApplicationAdapter {
             return false;
         }
     }
-    
+
 }
